Fixed postgres returned data type

+++ dina-test-support/src/test/java/ca/gc/aafc/dina/testsupport/ExampleDbBackedIntegrationTest.java
@@ -2,8 +2,6 @@ package ca.gc.aafc.dina.testsupport;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.math.BigInteger;
-
 import jakarta.inject.Inject;
 import jakarta.persistence.EntityManager;
 
@@ -25,7 +23,7 @@ public class ExampleDbBackedIntegrationTest {
         .createNativeQuery("select count(*) from pg_catalog.pg_type where typname = 'int2';")
         .getSingleResult();
 
-    assertEquals(BigInteger.valueOf(1), oneCount);
+    assertEquals(1L, oneCount);
   }
 
 }
